requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
nba_api
python-dotenv
pyodbc
